requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Flask-Migrate==4.0.5
pytz==2025.2
Werkzeug==2.3.3
DateTime